Release python-openstackclient 7.2.1
Recently we removed the 'volumev3' endpoint for cinder with change[1].
This caused issue with the openstackclient code relying on the
'volumev3' endpoint to check if cinder service is available or not
and was fixed by[2].
The above changes had an impact on the devstack code that set the
'volumes' quota for 'service' project to 50[3] (which was done to fix
some scenarios that started breaking now[4][5]) which stopped working
and we saw quota related failures in the glance-multistore-cinder-import job.

Oct 28 07:43:24.695646 np0038901070 devstack@g-api.service[88398]:
ERROR glance.api.v2.image_data [None req-1628d391-77e2-4714-9068-a080108bd41e
tempest-ImagesFormatTest-476245403 tempest-ImagesFormatTest-476245403-project-member]
Failed to upload image data due to internal error: glance_store.exceptions.BackendException:
VolumeLimitExceeded: Maximum number of volumes allowed (10)
exceeded for quota 'volumes'. (HTTP 413) (Request-ID: req-e8f6ced5-b38b-489d-a3ab-3dd9656ba9f1)

The purpose of this release is to apply the code fix[2] to gate jobs
and fix the gate.

[1] https://review.opendev.org/c/openstack/devstack/+/930010
[2] https://review.opendev.org/c/openstack/python-openstackclient/+/932640
[3] https://github.com/openstack/devstack/blob/2f3440dcfe27d760daa73d8a14d5bdbcf636a655/lib/cinder#L706-L709
[4] https://review.opendev.org/c/openstack/glance/+/932701
[5] https://zuul.opendev.org/t/openstack/build/c5c3931a3ea24861983025d022253367

Change-Id: I6c504329aaf3be63141913ea07047e4f70f43ded