[blazar-tempest-plugin] Tag 2023.1-last
2023.1 branch is moving to unmaintained state now[1]. Tempest and
tempest plugins are branchless which means master version of
Tempest and its plugins are not guaranteed to support unmaintained
branched. To continue testing the unmaintained branches, Tempest
and its plugins compatible tag[2][3] needs to be released for
their testing.

Tagging '2023.1-last' with the same hash of new tag 0.15.0
which mark the end of support for stable/2023.1 in Tempest.

[1] https://review.opendev.org/q/topic:%22zed-unmaintained%22
[2] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html
[3] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Change-Id: Ie76cc7bd3a50a6d5f3d20e39c1c09c9cc1d7e646