Release oslo.config for Epoxy-1 milestone
This is a library release for oslo.config for the Epoxy-1 milestone.
This repository contains commits that have not been released for this
cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

Change-Id: Ibba720a3ef5636ba38604ecf76d4c1bb5e03c1f8
Signed-off-by: Elod Illes <elod.illes@est.tech>