Release oslo.messaging for Epoxy-1 milestone
This is a library release for oslo.messaging for the Epoxy-1 milestone.
This repository contains commits that have not been released for this
cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

NOTE(elod.illes): MAJOR version bump is due to patch:
7997a199 Remove deprecated AMQP1 driver

Change-Id: I399bc89f2eda75da08fb88fe79453ca598b33db1
Signed-off-by: Elod Illes <elod.illes@est.tech>