Transition wallaby-eom branches to EOL
These projects did not even merge the bot-generated updates
to the .gitreview file for the unmaintained/zed branch, so obviously
nobody is caring for them and they should transition to EOL now that
the next branch (2023.1) has reached the end of maintainance.

Since EOL of unmaintained/zed means that all older branches need to be
eoled, too, this patch deals with wallaby first.

Change-Id: I00e8dd8d048cede42dc56e0c9d462979f7c01407