Update tooling to delete unmaintained branches
The tools/list_eol_stale_branches.sh script needs to handle
unmaintained/* branches in addition to stable/* branches, the former
also need to be deleted once the matching *-eol tag has been added to a
repository. In addition it needs to handle the delta between series and
branch names starting with 2023.1/antelope.

Also add a script to abandon open reviews on unmaintained branches, which
is a requirement in order to be able to delete them.

Change-Id: Iba0a50ad0d92013362e2a5eca5a4b1968e6acaf9