Mark murano-pkg-check deliverable abandoned
A governance/deliverable consistency check revealed that independent
deliverable murano-pkg-check is still defined in the releases repository
while no longer being defined in the governance repository.

Independent deliverables need to have "release-model: abandoned" in
order to be ignored in future checks while keeping their release history
intact. This is what was added to this one.

Change-Id: I10f01f763f8748d270bb20f5111a90a32d42dc99