Release PBR 6.1.1.0b1
Beta-test for PBR 6.1.1 because its underlying packaging model is
changing from traditional setup.py-based to PEP 517 pyproject.toml
build backend.

This should have absolutely no user-facing impact, as it's only
relevant to the process of building sdists of PBR itself, so at
worst could break our own releasing of PBR. Nevertheless, out of an
abundance of caution we're starting with a beta pre-release and will
then exercise it in another project before we follow up with a
proper release (or yank the beta from PyPI if it causes
unanticipated disruption).

Change-Id: I4761bfbc4a949bda0754e3b2204eb3d0dc8720d0