Release mistral-lib for Epoxy-2 milestone
This is the Epoxy-2 milestone release for mistral-lib.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If the team is happy with the current release hash, please +1 this
patch and we will process the release. If you need to get some
important changes merged first, leave a -1 and update the patch with
a new commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed after January 9th.

Change-Id: Ic4a1912a2a4aec4bf6b6a56d7274c98728978806
Signed-off-by: Elod Illes <elod.illes@est.tech>