Release python-novaclient for Epoxy-2 milestone
This is the Epoxy-2 milestone release for python-novaclient.
Clients and libraries following the cycle-with-intermediary release
model are encouraged to release each milestone to make sure updates
are made available.

If the team is happy with the current release hash, please +1 this
patch and we will process the release. If you need to get some
important changes merged first, leave a -1 and update the patch with
a new commit hash to use once it is ready.

If a release is definitely not wanted at this time, please -1 and
leave a comment explaining that and we can abandon the patch.

Any patches with no response will be assumed to be OK and will be
processed after January 9th.

Change-Id: I5c5d25c4d5191d77b4b9b69e37c97b2f32a14d06
Signed-off-by: Elod Illes <elod.illes@est.tech>