Releasing release-test
...to test the zuul-jobs change from direct invocation of setup.py
to using the pyproject-build utility. This is going to wind up using
a newer SetupTools version, which will normalize project names in
sdist filenames similar to how wheel filenames are already
normalized, so we want an end-to-end test to see where problems
arise. The Depends-On is the zuul-jobs change in question, so that
this can't merge until that does.

While we're here, clean up a stray comment line that got duplicated
from the dalmatian file when epoxy was created.

Depends-On: https://review.opendev.org/940273
Change-Id: I626c3e3c5905b5b6447841fe05fd2d3ad4934450