Release bifrost for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to bifrost since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 17.1.0..f3f879c7
f3f879c7 Fix Keystone default log location typo
7dff12e5 Fix key-order ansible errors
d45ddf8a ironic: Perform online data migrations with localhost DB
0664aeed Fix BRANCH for install-deps.sh
e9d024ff Switch testing to stable/2023.2
8cf671f4 Update TOX_CONSTRAINTS_FILE for stable/2023.2
b62d5d27 Update .gitreview for stable/2023.2

Change-Id: I665b7ecf720c8af0545936cd04da8c68bdef30dc
Signed-off-by: Elod Illes <elod.illes@est.tech>