Release ceilometer for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to ceilometer since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 21.0.0..6222e6a08
6222e6a08 devstack: Use spaces instead of tabs
74e7e6be1 Add sg-core backend support in devstack plugin
c02b67723 Fallback to oslo_cache.dict caching backend
8961c1d19 Update TOX_CONSTRAINTS_FILE for stable/2023.2
fca9a772e Update .gitreview for stable/2023.2

Change-Id: Ia5501dcecbab494a7bf92cff2dafd130d7c04fbe
Signed-off-by: Elod Illes <elod.illes@est.tech>