Release glance for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to glance since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 27.1.0..18453f860
18453f860 Fixed typo in discovery.py
5b4be5b6f Add iso file format inspector
8540ffc6e Fix LP#2054575
3b3141baa Revert image state to queued if conversion fails
9e28810cc Do not set_acls if store is not associated to glance node
f6557db0a Fix /healthcheck url typo in glance-api-paste

Change-Id: Ie740f48a38bb605ce0225500985a9fc1ef83a8dc
Signed-off-by: Elod Illes <elod.illes@est.tech>