Release heat for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to heat since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 21.0.0..b6af90341
ed3069fed Stable-only: Drop grenade job
b92ddb294 Avoid wrongly passing "no_fixed_ips" to Neutron
4fd9953a5 Try archive path to download Fedora image
39eca5de2 Remove reference to devstack-gate
933f92189 Revert "Skip aodhclient plugin test"
4ca77ff93 Fix update of OS::Heat::UpdateWaitConditionHandle
d1363cc17 Update TOX_CONSTRAINTS_FILE for stable/2023.2
18d4ca055 Update .gitreview for stable/2023.2

Change-Id: I100e5badfc0e4fc46ecb431e8a2e1bdb9d1573ed
Signed-off-by: Elod Illes <elod.illes@est.tech>