Release horizon for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to horizon since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 23.3.0..c873490d3
c873490d3 Fixing Incorrect URL when browsing Swift containers
0208b3e1d Fix generation of inventory capacities
76e07beed Delete container show duplicate toast notifications
bca5380ef Fix Placement statistics display
d964bed17 Sort image source choices by name for volume
fdacbacec Fix floating IP associated to unbound port
481b1acf4 Fix allowed address pair row unique ID
f8ddd31e1 Show Created At column for backups table
569f59bce Show availability zone for volume backups
6f0e86358 Respect SSL settings in placement API
593ef9b56 Fix Users/Groups tab list when a domain context is set
e9bf69564 Fix error on changing user password by admin
4de36bb64 Extend configurable skippability of neutron calls to project instance detail
06e0a3e04 Imported Translations from Zanata
e65012797 Add allocation ratios to Placement stats
ac9b05011 bandit: Install baseline extra
034630c4d Use Placement API along with the hypervisor stats
3c6029cd9 Imported Translations from Zanata
204ca2876 Update TOX_CONSTRAINTS_FILE for stable/2023.2
6677e8c92 Update .gitreview for stable/2023.2

Change-Id: Ib23a847689b007ebae5d8e7516b85735c3db8532
Signed-off-by: Elod Illes <elod.illes@est.tech>