Release ironic-lib for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to ironic-lib since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 5.5.0..eff1475
eff1475 Fix invalid UTF-8 characters in execute output
8d37c00 Update TOX_CONSTRAINTS_FILE for stable/2023.2
d975cfb Update .gitreview for stable/2023.2

Change-Id: Idaa0e3971aac62b85cd7087269e7f229375987f3
Signed-off-by: Elod Illes <elod.illes@est.tech>