Release ironic-python-agent for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to ironic-python-agent since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 9.7.2..72cba7
83363b1f [CI] Pin sushy-tools
8dbbe793 Fix errors in the function erase_devices_express
47f60559 Fix RAID volume name
6a32f442 Remove non RE2 job config

Change-Id: I393315b1e7a58a9f5f8ae5a6c9cffa70912472a7
Signed-off-by: Elod Illes <elod.illes@est.tech>