Release ironic for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to ironic since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 23.0.3..a3b014965
a3b014965 Fix redfish session cache on missing password
6593ec86b [CI] Pin sushy-tools
5efbf9a22 Calculate missing checksum for file:// based images
9925ed5f4 Fix actual size calculation for storage fallback logic

Change-Id: I009c1b702b3de5950bbf09178ed9daa5cb687bfd
Signed-off-by: Elod Illes <elod.illes@est.tech>