Release keystone for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to keystone since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 24.0.0..1829ce020
1829ce020 Add keystone-manage reset_last_active command
fee954e9b Remove reference to devstack-gate
7697140fc Add domain scoping to list_domains
f519bcedf Allow domain admin to view roles
4711b36ab Allow domain users to manage credentials
570c19e91 Allow admin to access tokens and credentials
2abaaae89 Add ability to create users and projects from keystone-manage
81dbfd265 Remove unused old job templates and experimental jobs
932863b6d Improve application credential validation speed
5a55e9de1 Normalize policy checks for domain-scoped tokens
7de9f7d49 Allow users with "admin" role to get projects
62858870b Fix policies for groups
991662c66 Consistent and Secure RBAC (Phase 1)
653d82b1b Update TOX_CONSTRAINTS_FILE for stable/2023.2
c0662c87b Update .gitreview for stable/2023.2

Change-Id: I224532e844e9b74508392e39c012b00012f9ac4d
Signed-off-by: Elod Illes <elod.illes@est.tech>