Release keystoneauth for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to keystoneauth since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 5.3.0..5c38731
5c38731 Fix "dictionary changed size during iteration"
6f4b4ba Pass oidc client id in the body if no client secret provided
565c6c0 Update TOX_CONSTRAINTS_FILE for stable/2023.2
59f063d Update .gitreview for stable/2023.2

Change-Id: I044a9b21e76cd56b3e63d0a1ea24bb5ff3a46f53
Signed-off-by: Elod Illes <elod.illes@est.tech>