Release networking-generic-switch for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to networking-generic-switch since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 7.2.1..4d320c5
4d320c5 Do not make actual device changes in bind_port()
b23d0a7 [CI] Remove not defined job from queue

Change-Id: I06d382dfe951332447a6ccedc605bdb3b7d89a84
Signed-off-by: Elod Illes <elod.illes@est.tech>