Release neutron-fwaas for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to neutron-fwaas since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 19.0.0..e838bd86c
e838bd86c create default security group using an independent context
da92be122 [stable 2023.2 only] CI fixes
74f0f655f Update TOX_CONSTRAINTS_FILE for stable/2023.2
994e9c0a1 Update .gitreview for stable/2023.2

Change-Id: Ia9c22c43d16bcec50c66ba882e0de945b976c033
Signed-off-by: Elod Illes <elod.illes@est.tech>