Release neutron for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to neutron since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 23.3.0..d1c69571c8
d1c69571c8 OVS: Set log level to INFO for unconfigured ofport cases
dfe3dc21e0 [OVN] Set reside-on-chassis-redirect also for FLAT networks
137f2c4f59 [OVN] Add the network type to the ``Logical_Switch`` register
2f6507e383 Clean up state VRRP PID file
527ca52880 [Functional tests] Add logging router interfaces in metadata IPv6 tests
97a1003b36 functional: Handle ovsdb monitor returning inserts in different checks
c933c287cf ovn: Disable activation-strategy=rarp for DPDK ports
03bf0d7c38 Make sure that policy enforcer is initialized before use
0e3419c620 Allow network owner reader to get subnets
81e38bc897 Use the correct input for OVN agent deletion
f30d7c7093 Revert "[HA] Do not add initial state change delay in HA router"

Change-Id: I507cbc2a13f625ff5c36531ff4ab05ab6ddd6bb7
Signed-off-by: Elod Illes <elod.illes@est.tech>