Release openstacksdk for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to openstacksdk since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 1.5.0..30a7bfc3a
a4dd3d7d8 [Stable Only] Do not pin openstacksdk to master
a36387e53 compute: Add locked filter for Server
8e9e4cc97 Remove retired project Senlin job
090ca81fa Fix list of server migrations
5fce090ea [baremetal] Add support for querying for shards
57ff50d47 Update TOX_CONSTRAINTS_FILE for stable/2023.2
7058ddddf Update .gitreview for stable/2023.2

Change-Id: Id0a697a614e8566081431eb8e701e96260e4d2a0
Signed-off-by: Elod Illes <elod.illes@est.tech>