Release oslo.db for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to oslo.db since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 14.1.0..8fb36e9
8fb36e9 exc_filters: Handle OperationalError for MariaDB/Galera
e8c9c9b Remove tips job
e0ee12f bindep: Use new mysql-* package names

Change-Id: I62a84ae1e9604be06ac13773e6617500f41b6449
Signed-off-by: Elod Illes <elod.illes@est.tech>