Release oslo.log for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to oslo.log since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 5.3.1..6c08a35
6c08a35 Apply eventlet workaround only once
573a434 Fix broken reference to rate_limit_except_level

Change-Id: Iec7d762d29972478bd4e7207b8e3319900dba0df
Signed-off-by: Elod Illes <elod.illes@est.tech>