Release python-manilaclient for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to python-manilaclient since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 4.6.1..815c0dd
815c0dd [OSC] Fix output format for osc share subnet create
d8d7acf Don't set availability_zone for 2.32 version of api
d7193cd Fix list call when search_opts isn't provided
6f53844 Fix share force delete case
2614b30 Fix "per_share_gigabytes" in "openstack quota set"

Change-Id: Iba3c3fb55899486e27e2055979c71d0c32a31ecb
Signed-off-by: Elod Illes <elod.illes@est.tech>