Release python-vitrageclient for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to python-vitrageclient since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 4.8.0..d5ab44e
d5ab44e Support networkx 3.x
8a8defc Update TOX_CONSTRAINTS_FILE for stable/2023.2
19226bd Update .gitreview for stable/2023.2

Change-Id: I980557a77b4dc34fc7d1d64d61667cd442db375a
Signed-off-by: Elod Illes <elod.illes@est.tech>