Release swift for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to swift since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 2.32.0..1cc23d54b
1cc23d54b docs: Call out that xprofile is not intended for production
0236ddaef xprofile: Stop using eval()
b47d0cf53 Ensure correct content-type in container HEAD response
0de976cf5 [stable-only] CI: Remove some old rolling-upgrade jobs
0ad40ea1a CI: Move off CentOS 8
e39fd8b84 CI: fix rolling-upgrade jobs
ce1e641e3 CI: Remove centos-7 jobs
d0ce5821c debug_logger: ensure NOTICE level is registered
e01b28b48 Update TOX_CONSTRAINTS_FILE for stable/2023.2
468306d4f Update .gitreview for stable/2023.2

Change-Id: I3d4815644295f9a1aea00972a81601f4c4f6d770
Signed-off-by: Elod Illes <elod.illes@est.tech>