Release watcher for stable/2023.2
As a preparation for the transition of 2023.2 Bobcat to End of Life [1],
this release picks up new commits to watcher since
the last release from stable/2023.2.

This is being proposed as a convenience to help make sure stable
changes are being released. If the team is good with this going out,
please respond with a +1 to let the release team know it is OK to
proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 11.0.0..090b6959
3874ca82 Unblock the CI gate
b5431b2e Handle deprecated "cpu_util" metric
c3640ed6 Update TOX_CONSTRAINTS_FILE for stable/2023.2
1f8566a9 Update .gitreview for stable/2023.2

Change-Id: I987606673d8d4ee1deda0505426cf5065da9bb59
Signed-off-by: Elod Illes <elod.illes@est.tech>