Release final oslo.config for 2025.1 Epoxy
This proposed a release for oslo.config to get any unreleased
changes out prior to library freeze for 2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 20th.

Change-Id: I1df186b6aaccc43d46170df2ccae63cb2e3770f8
Signed-off-by: Elod Illes <elod.illes@est.tech>