Release final oslo.context for 2025.1 Epoxy
This proposed a release for oslo.context to get any unreleased
changes out prior to library freeze for 2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 20th.

Change-Id: Ife03f67a8e0f45b92af85daa5eda5c034e8eb94c
Signed-off-by: Elod Illes <elod.illes@est.tech>