Release final oslo.db for 2025.1 Epoxy
This proposed a release for oslo.db to get any unreleased
changes out prior to library freeze for 2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 20th.

Change-Id: Iac6f34aa9d0ecd3b3fd5020bfe9c8be5b378287b
Signed-off-by: Elod Illes <elod.illes@est.tech>