Release final oslo.reports for 2025.1 Epoxy
This proposed a release for oslo.reports to get any unreleased
changes out prior to library freeze for 2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 20th.

Change-Id: I7bbc407c191bc79828a883b98fb4c18fafc9d2b0
Signed-off-by: Elod Illes <elod.illes@est.tech>