Release final oslo.versionedobjects for 2025.1 Epoxy
This proposed a release for oslo.versionedobjects to get any unreleased
changes out prior to library freeze for 2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 20th.

Change-Id: I019cba1681d7268a3d8420c99c2da132d8419ff7
Signed-off-by: Elod Illes <elod.illes@est.tech>