Release final stevedore for 2025.1 Epoxy
This proposed a release for stevedore to get any unreleased
changes out prior to library freeze for 2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready.

Patches with no response will be merged after Thursday, February 20th.

Change-Id: Ia16bd36c80949adad4aba8d9d5591c2c9a5f3bc4
Signed-off-by: Elod Illes <elod.illes@est.tech>