Transition unmaintained/xena to EOL
After the 2023.1 series has been transitioned to unmaintained, all older
unmaintained branches will be moved to EOL according to the policy[0].

If you want to opt-in to keep some of these branches in the unmaintained
status, please note the following:

- You need to opt-in to keep all branches from xena through to
  2023.1 active, there are no gaps allowed.
- The CI for those branches must be in good shape, this implies in
  particular:
  - All bot-generated patches like updates to the .gitreview config are
    merged.
  - There are no zuul config errors for these branches[1].
  - All configured zuul jobs are passing on a test patch.

If all of these conditions are met, you can submit a review on top of
this one which reverts the creation of the xena-eol tags for the
repos that you opt into. Please link to the test patches showing the
working CI. Once verification of the conditions is completed, your
revert will be applied to this change.

The deadline for the opt-in process will be 21 Mar 2025, after that this
change will be merged. The whole process can also be tracked at [2].

[0] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html
[1] https://zuul.opendev.org/t/openstack/config-errors?branch=unmaintained%2Fxena&skip=0
[2] https://etherpad.opendev.org/p/eol-unmaintained-branches

Change-Id: Ia19103ad21c4ea71af6b50972f82154bd7ced6f3