Release final python-glanceclient for 2025.1 Epoxy
This is a release for python-glanceclient to get all the merged but
unreleased changes out prior to client library freeze for
2025.1 Epoxy going into effect.

It is important that if the team is ready to proceed, then please leave
a +1 to acknowledge this, so that we can process this right away.
If the team is working on any final merges, leave a -1 with some detail
on the status and update the patch with the new commit hash to use when
ready. Either way, please do respond on this patch.

In any case, if the patch got no response, then this will be merged
after Thursday, February 27th.

Change-Id: Ie5e265b271112f1bb4a2cd7f320a0f19e49c8d08
Signed-off-by: Elod Illes <elod.illes@est.tech>