Release networking-sfc RC1 for 2025.1 Epoxy
This proposed a RC1 release for networking-sfc to get any merged changes
out prior to the RC1 deadline for 2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready. If there is nothing
to wait for, then please do respond on this patch with +1 as soon as
possible.

In any case, if the patch got no response, then this will be merged
on Friday, March 14th.

Change-Id: I5d4bf8717d8aa7c58c086082d1d50be15814acda
Signed-off-by: Elod Illes <elod.illes@est.tech>