Release barbican-tempest-plugin for 2025.1 Epoxy
This patch proposes a release for barbican-tempest-plugin
to get the latest merged changes out prior to the final
2025.1 Epoxy going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready. If there is nothing
to wait for, then please do respond on this patch with +1 as soon as
possible.

In any case, if the patch gets no response, then this will be merged
after Thursday, March 20th.

Change-Id: I6bf063815dd7abc764bf7ed7755554ebc4c9bbe2
Signed-off-by: Elod Illes <elod.illes@est.tech>