[horizon] Cut stable/2025.1 branch
This patch cuts missing stable/2025.1 branches for deliverables that are
missing it.

We are in Feature Freeze period already, but if the team needs to do a
another final release because of any release critical fixes, then please
indicate it with a -1.
If there is nothing to wait for, then please do respond on this patch
with +1.

In any case, if the patch gets no response, then this will be merged
after Thursday, March 20th.

Change-Id: I5a42c8936e5942187813ef2bfa7f2660431c001f