[process] Update www landing page refresh task
The task is outdated as it still refers to Documentation team that does
not exist anymore. This task was long time discussed and added to TC's
duties to take care of.

The task description is now updated with refering to TC and added a
note with a link of detailed task description in Documentation
Contributor Guide, and also some link to example patches.

Change-Id: I3da8123de9aa5406f8cbc82882ef17831ce7158b