[tool] Fix unreleased-since listing
'list-deliverables --unreleased-since <date>' command gave false
results as it did not list deliverables based on the release date,
rather the date of the commit that was tagged.

The cause of the false behaviour is that the wrong api call was used:
it got the Commit object based on the tag's sha (ID), instead of the
TagInfo object.

Change-Id: I530ad335b998a809e2dbaea5a03f460120962dd8