[Puppet OpenStack] Transition 2023.2 Bobcat to End of Life
With the new resolution '2023-07-24 Unmaintained status replaces
Extended Maintenance' [1] non-SLURP releases (starting with 2023.2
Bobcat) are not eligible to move to Unmaintained after their maintained
phase. Hence, with this patch 2023.2 Bobcat moves to End of Life.

Teams had time to prepare final 2023.2 Bobcat stable releases already,
but in case any of the deliverables needs another release (for example
to fix a regression) before the transition, then please add a -1 for
this patch and propose a release patch. If the team is OK with the
transition then please do add a +1.

The deadline is (for non-trailing projects) April 30th, 2025.

[1] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html

Change-Id: I6d83c45b64d17ca904ac864049ca80b92f0fa598