Release oslo.service for Flamingo milestone
This release includes the new threading backend in oslo.service.

If the team is ready to proceed with this release, please leave a +1
to indicate we should go ahead.

If more time is needed for pending changes, or if there are other
reasons to delay the release, please leave a -1 with a comment
explaining the current status. Once ready, this patch can be updated
with the new commit hash to proceed with the release.

Change-Id: Ie4603517edd610f7fd65d8629313e7a1c8644a30