Release oslo.cache for Flamingo milestone.
If the team is ready to proceed with this release, please leave a +1
to indicate we should go ahead.

If more time is needed for pending changes, or if there are other
reasons to delay the release, please leave a -1 with a comment
explaining the current status. Once ready, this patch can be updated
with the new commit hash to proceed with the release.

Change-Id: I5611569869b264ac50da35177a8f27700cf418e9