Comply with Policy for AI Generated Content
While probably not meeting most people's definition of artificial
intelligence these scripts do create patches. The OpenInfra Policy
for AI Generated Content recommends to:

> Add a "Generated-By:" label to the commit message, and explain in
> comments or the commit message any prompts or background context
> the reviewers might need to fully understand the change and how
> much of the change was generated by the tool.

Even though it also has a carve-out for these sorts of scripts:

> The exception to the rule is that we do allow submissions from
> well documented automated processes, such as release tooling or
> for internationalization updates.

https://openinfra.org/legal/ai-policy

Nevertheless, documenting within the commit message the location of
the script which was used to generate that contribution is helpful
for a number of obvious reasons. Let's do it anyway, even though
it's not technically required.

Change-Id: Id798bba638543cf8e9573c10e85495e7cc436237