[tempest] Tag 2023.2-last
2023.2 release is EOL. Tempest and tempest plugins are branchless
which means master version of Tempest and its plugins are not
guaranteed to support unmaintained branched. To continue testing
the unmaintained branches, Tempest and its plugins compatible
tag[2][3] needs to be released for their testing.

Tagging '2023.2-last' with the same hash of new tag 44.0.0
which mark the end of support for stable/2023.2 in Tempest.

[1] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html
[2] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Change-Id: Ia6e847f5a93515a8390a754e6429ec9660568db9