Release designate for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to designate since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 18.0.0..af7445fa
af7445fa Add an designate-api-wsgi compatibility script
a33f7376 Support "-" in region name for FIP's
b905f44d api: Handle duplicated query string paramters
7e58bdd5 Fix YAMLAdapter when optional attributes are missing
50b04e63 Stable-only: Switch neutron job to stable version
d60cba7a devstack: Fix manage.py execution with GLOBAL_VENV
72777e20 Update TOX_CONSTRAINTS_FILE for stable/2024.1
3cd69a20 Update .gitreview for stable/2024.1

Change-Id: I92066e0eef1500c36092becf121697ae555fe6d9
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh