Release ironic-lib for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to ironic-lib since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 6.0.0..aa4a2cd
aa4a2cd Fix invalid UTF-8 characters in execute output
9be44cf Update TOX_CONSTRAINTS_FILE for stable/2024.1
7b9c14e Update .gitreview for stable/2024.1

Change-Id: I55b4ed4c626af62b5ec15057fd6b0f93835ac74d
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh