Release metalsmith for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to metalsmith since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 2.1.1..b07b2e7
7ebd5c9 Use bootloader element instead of grub2
dcaa1f6 Update TOX_CONSTRAINTS_FILE for stable/2024.1
269a05c Update .gitreview for stable/2024.1

Change-Id: I69319195b1648a32d5bea1473c548c0da0e0f9b7
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh