Release networking-baremetal for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to networking-baremetal since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 6.3.0..8dc6eda
8dc6eda prevent break on communications failure
01203f3 avoid attribute error on bad password or config
62ca32e Update TOX_CONSTRAINTS_FILE for stable/2024.1
632c03d Update .gitreview for stable/2024.1

Change-Id: I2b9b685145299772a55fd6114c01b7e33c06501b
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh