Release neutron-fwaas for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to neutron-fwaas since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 20.0.0..157f275f6
157f275f6 create default security group using an independent context
0d3bea455 [stable-only] Switched to branched tempest plugin job
3fa0d6902 Remove devstack-gate requirement
6c5a0536c Update TOX_CONSTRAINTS_FILE for stable/2024.1
8112ca3fa Update .gitreview for stable/2024.1

Change-Id: I5c1e7a761dd37f6299920b5c108a98a373646e2f
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh