Release neutron-vpnaas for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to neutron-vpnaas since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 24.0.1..8655ef2b0
8655ef2b0 Use write connection for get_vpn_services_on_host
2dbbc16f3 Stable Only: Use branched tempest jobs
cf4497fec Remove reference to devstack-gate

Change-Id: Id9889d13e51a0f48d22c8c87971e93bde09ff4dc
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh