Release neutron for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to neutron since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 24.2.0..24c492cf39
24c492cf39 Limit trunk ACTIVE state hack to OVN
6634ef5110 Consider logging options when using OVNdbsync
d363b51503 Improve ACL comparison efficiency
e7713e8742 Allow empty gateway IP in subnets from subnet pools
7b0a44c02b [FT] Improve ``test_assert_pings_during_br_phys_setup_not_lost*`` tests
4128a867c1 Fix nova notification on port unplug
bcc380b39e Add provider network resources filter for "network"
1814fbb4c8 [OVN] OVN agent retrieval filter matching improvement
790b51985a Allow to delete singleton instances
420b5dc22f [FT] Match the correct event in ``test_subport_delete``

Change-Id: Iccc2635cec417d143b9f3c4d1b0abaa1b0af3912
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh