Release nova for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to nova since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 29.2.1..ad1bd175ae
ad1bd175ae Fix detaching devices by alias with mdevs
4d01fb17e6 Add repoducer test for bug 2074219
db438e55e6 [tool] Fix backport validator for non-SLURP
0b738435c9 live migration: Avoid volume rollback mismatches
d655d98350 Amend functional reproducer for bug 1899835

Change-Id: I41237ee706e80ec7fc17bf79bf97f50f1a6ad699
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh