Release os-ken for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to os-ken since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 2.8.1..073d6f05
073d6f05 Raise ValueError in case unpack_from returns zero length
06b08e77 Update TOX_CONSTRAINTS_FILE for stable/2024.1
aef25468 Update .gitreview for stable/2024.1

Change-Id: I8dc50e61cfd9a3389dac8a6a10c2181f9abc0cf1
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh