Release oslo.cache for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to oslo.cache since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 3.7.0..5d7c471
5d7c471 Use BMemcacheClientPool if tls_enabled=true
5f0f768 Update TOX_CONSTRAINTS_FILE for stable/2024.1
6f5190c Update .gitreview for stable/2024.1

Change-Id: Iedbcf0e3e02e1f516364ec535e64e7ac818df378
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh