Release ovn-octavia-provider for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to ovn-octavia-provider since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 6.0.0..43bbc33
43bbc33 [stable-only] Update stable/2024.1 tempest jobs
686795a Remove join on helper request daemon thread
9a182ec Fix pep8 with pylint 3.3.0
1716619 Fix member subnet id on a fully populated LB
edd1876 Error log for missing certs with NB and SB DBs
8cabfc5 Maintenance task: do not change IPv4 ip_port_mappings
2ced829 [stable/2024.1 only] Drop -master jobs
4a1c630 Don't create an OVSDB connection per API request
4df7858 Return DELETED status when deleting a nonexistent member
c14eb46 Reduce coverage threshold on stable branches
e806547 Remove leftover OVN LB HM port upon deletion of a member
de21f35 FIX OVN LB Health Monitor checks for IPv6 members
c68129c Adding isolation to functional tests
d48cefd Fix new pylint errors
6816d67 Update TOX_CONSTRAINTS_FILE for stable/2024.1
c1e9124 Update .gitreview for stable/2024.1

Change-Id: Ia0e0486caee57d32852778e9fb59fb5dc12e3476
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh