Release python-manilaclient for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to python-manilaclient since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 4.8.0..453dc27
453dc27 Replace OS_TENANT_NAME
2db7610 Don't set availability_zone for 2.32 version of api
e5443d8 [OSC] Fix output format for osc share subnet create
f8ea4a7 Update TOX_CONSTRAINTS_FILE for stable/2024.1
0bac0dc Update .gitreview for stable/2024.1

Change-Id: I44c3572cc9021f41e908f6af089e28924651d819
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh