Release python-octaviaclient for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to python-octaviaclient since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 3.7.0..b7a9d91
b7a9d91 Fix displaying list columns with null values
eb8c0e7 osc: Use FormattableColumn for formatter
f24c0b9 Update TOX_CONSTRAINTS_FILE for stable/2024.1
242c6d8 Update .gitreview for stable/2024.1

Change-Id: Ib263a3b639aca62e042de65d0de81e4e6fb9d13e
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh