Release skyline-apiserver for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to skyline-apiserver since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 4.0.1..718cfe2
718cfe2 Change to use shell cat to fetch commit_id
9b85645 fix: Set stable-2024.1 version of skyline-console to use
c7195d5 Fix skyline-apiserver on python3.11
b18edc1 Update TOX_CONSTRAINTS_FILE for stable/2024.1
c5a20b9 Update .gitreview for stable/2024.1

Change-Id: I46f8264a7db8edf6839430ebec37e7918ebe66bc
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh