Release swift for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to swift since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 2.33.0..fac2e759b
fac2e759b docs: Call out that xprofile is not intended for production
18b0df30b xprofile: Stop using eval()
e0818f55b Ensure correct content-type in container HEAD response
85d839784 [stable-only] CI: Remove some old rolling-upgrade jobs
4a89cafc1 CI: Move off CentOS 8
46aabac09 Update TOX_CONSTRAINTS_FILE for stable/2024.1
24633bda2 Update .gitreview for stable/2024.1

Change-Id: I87fa3167b498f918ebf2879c24b67f3b1b7219d1
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh