Release taskflow for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to taskflow since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 5.6.0..24c88865
24c88865 Fix REVERT_ALL with Retries in unordered Flows
ce81b920 Remove SQLAlchemy tips jobs
34ef1ea9 Add functional job with redis enabled
2b7f58cf Fix broken unit tests for redis jobboard driver

Change-Id: I44d3ab16bd8c3d65f7878b9e462c0a5a6550f2ea
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh