Release trove for stable/2024.1
As a preparation for the transition of 2024.1 Caracal to Unmaintained
[1], this release picks up new commits to trove since
the last release from stable/2024.1.

This is being proposed as a convenience to make sure that merged stable
changes are released. If the team is good with this going out, please
respond with a +1 to let the release team know that it is OK to proceed.

If it is not wanted at this time, or if there are more changes that
would be good to get merged before doing a stable release, please
leave a -1 with a comment with what the team would prefer. We can
then either abandon this patch, or wait for an update with a new
commit hash to use instead.

[1] https://releases.openstack.org/

$ git log --oneline --no-merges 21.0.1..25b7d92a
25b7d92a [CI]: run perodical jobs on master branch only
69d7203a Fix MariaDB target dir for backup syntax
b7363a27 Enable network_isolation by default
87d60a1b Change settings for stable/2024.1
d2877ee3 fix error when build mariadb backup docker image
452a9c27 Fix error can't detach the configuration group binding
89f5bc36 Get rid of nose and Proboscis

Change-Id: Ifd6b4bb5c00f4e115ad60c166e0167cb397b5d64
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh