Release final oslo.log for 2025.2 Flamingo
This proposed a release for oslo.log to get any unreleased
changes out prior to library freeze for 2025.2 Flamingo going into
effect.

It is important that if the team is ready to proceed, then please leave
a +1 to acknowledge this, so that we can process this right away.
If the team is working on any final merges, leave a -1 with some detail
on the status and update the patch with the new commit hash to use when
ready. Either way, please do respond on this patch.

In any case, if the patch got no response, then this will be merged
after Thursday, August 21st.

Change-Id: I0a022adc9f36d3bf57e368e3c4eb7c294638cd97
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh