Release final python-brick-cinderclient-ext for 2025.2 Flamingo
This is a release for python-brick-cinderclient-ext to get all the merged but
unreleased changes out prior to client library freeze for
2025.2 Flamingo going into effect.

It is important that if the team is ready to proceed, then please leave
a +1 to acknowledge this, so that we can process this right away.
If the team is working on any final merges, leave a -1 with some detail
on the status and update the patch with the new commit hash to use when
ready. Either way, please do respond on this patch.

In any case, if the patch got no response, then this will be merged
after Thursday, August 28th.

Change-Id: Id978356081c24071481a49130c0ede424d5f6752
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh