Release final python-watcherclient for 2025.2 Flamingo
This is a release for python-watcherclient to get all the merged but
unreleased changes out prior to client library freeze for
2025.2 Flamingo going into effect.

It is important that if the team is ready to proceed, then please leave
a +1 to acknowledge this, so that we can process this right away.
If the team is working on any final merges, leave a -1 with some detail
on the status and update the patch with the new commit hash to use when
ready. Either way, please do respond on this patch.

In any case, if the patch got no response, then this will be merged
after Thursday, August 28th.

Change-Id: I87d811fa7801783113d872e7fa4743e9250a4378
Signed-off-by: Elod Illes <elod.illes@est.tech>
Signed-off-by: Sean Mooney <work@seanmooney.info>
Generated-By: openstack/releases:tools/process_auto_releases.sh