[blazar] Create 2025.2 branch for client and non-client libs
This creates the stable/2025.2 branch for all client and non-client
libraries now that we are past the library freeze.

If the team wants to wait for a specific patch to make it to the
library (for example if the gate of the deliverable is broken and
needs to be fixed before branching), then someone from the team can
-1 the patch to have it held, or update that patch with a different
commit SHA. If there is nothing to wait for, then please do respond
on this patch with +1.

In any case, if the patch gets no response, then this will be merged
after Thursday, September 4th.

Change-Id: I916b75a03fe6fd8f52a327a8ce6887eaf9f6f22f
Signed-off-by: Előd Illés <elod.illes@est.tech>
Generated-By: openstack/releases:tools/bulk_review.sh