Release zaqar RC1 for 2025.2 Flamingo
This proposed a RC1 release for zaqar to get any merged changes
out prior to the RC1 deadline for 2025.2 Flamingo going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready. If there is nothing
to wait for, then please do respond on this patch with +1 as soon as
possible.

In any case, if the patch gets no response, then this will be merged
on Friday, September 12th.

Change-Id: Ie0b1c6a5ab28e76e57588c5c539833d820b8f606
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh