[horizon] Cut stable/2025.2 branch
This patch cuts missing stable/2025.2 branches for deliverables that are
missing it.

We are in Feature Freeze period already, but if the team needs to do
another final release because of any release critical fixes, then please
indicate it with a -1.
If there is nothing to wait for, then please do respond on this patch
with +1.

In any case, if the patch gets no response, then this will be merged
after Thursday, September 18th.

Change-Id: I3fd2028286de08247998260c3ab22ad1e8d796ae
Signed-off-by: Előd Illés <elod.illes@est.tech>
Generated-By: openstack/releases:tools/bulk_review.sh