Release freezer-tempest-plugin for 2025.2 Flamingo
This patch proposes a release for freezer-tempest-plugin
to get the latest merged changes out prior to the final
2025.2 Flamingo going into effect.

If the team is ready to proceed, please leave a +1 to acknowledge this
release and we will process right away. If the team is working on any
final merges, leave a -1 with some detail on the status and update the
patch with the new commit hash to use when ready. If there is nothing
to wait for, then please do respond on this patch with +1 as soon as
possible.

In any case, if the patch gets no response, then this will be merged
early next week.

Change-Id: I9501f76f138f14d4078b9ad6d5740de25bff4a1b
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh