2025.2 Flamingo final releases for cycle-with-rc projects
This tags the final release for all cycle-with-rc projects for 2025.2
Flamingo. While not required, having PTLs ACK this patch would be nice
to include that record in the merged metadata.

Release activity should now be frozen until this is processed on
October 1st. Any bugfixes merged to stable/2025.2 can be queued up and
ready to go as a follow on stable release after the coordinated release
date.

There are different diff-start values for each project. This is used
when generating release notes to denote the point where the
stable/2025.1 branch started to diverge from what is included on
stable/2025.2. For some, this will be the final 2025.1 Epoxy major
version, but if multiple RCs were done, this may show the initial RC1
release version. This is normal.

This patch will be merged on October 1st.

Change-Id: I364f4d4be1f65dfc7afb12c1d83ac8433ba01972
Signed-off-by: Előd Illés <elod.illes@est.tech>