[octavia] Transition 2024.1 Caracal to Unmaintained
This branch reached the end of its official stable maintained period,
hence this patch transition stable/2024.1 branch to Unmaintained. (For
details about Unmaintained state please see: [1][2])

The transitions is as follows:
- tip of the stable/2024.1 branch is tagged with 2024.1-eom tag
- unmaintained/2024.1 branch will be cut from 2024.1-eom tag
- stable/2024.1 branch will be deleted

Note:
- there will no longer be official releases off of 2024.1 branch after
  this transition
- backports can be pushed to unmaintained/2024.1 branch
- openstack-unmaintained-core (or <project>-unmaintained-core groups if
  such exists) have the rights to merge patches

Important: please +1 if the team is ready for us to proceed with this
transition, or -1 if there are any final 2024.1 Caracal stable release
with backports currently in flight that we should wait for. For the
latter case, please propose a final 2024.1 Caracal stable release
patch, and rebase this patch on top of that, with the latest commit
hash of stable/2024.1.

The planned transition deadline is October 24th, 2025 [3] last week,
so please approve this patch as soon as possible if the project is
ready for the transition.

[1] https://docs.openstack.org/project-team-guide/stable-branches.html#unmaintained
[2] https://governance.openstack.org/tc/resolutions/20230724-unmaintained-branches.html
[3] https://releases.openstack.org

Change-Id: I772943af4248e7142dbe2476a27856023a6b7cc7
Signed-off-by: Előd Illés <elod.illes@est.tech>
Generated-By: openstack/releases:tools/bulk_review.sh