pbr 7.0.2
Remove reliance on deprecated SetupTools functionality that is due
to be removed *very* soon. This could potentially impact users of
PBR's internals like pbr.core (which was removed), but we haven't
officially considered that a supported public API even though we
didn't mark the functions as private. All changes since the last
release are refactors, test-related, or fixing problems that will
become breaking bugs soon if we don't address them immediately.

Change-Id: I29cac66277deb12ff8bb67ce7c91cbf09e100a7f
Signed-off-by: Jeremy Stanley <fungi@yuggoth.org>