[zaqar-tempest-plugin] Tag 2024.1-last
2024.1 branch is moving to unmaintained state now[1]. Tempest and
tempest plugins are branchless which means master version of
Tempest and its plugins are not guaranteed to support unmaintained
branched. To continue testing the unmaintained branches, Tempest
and its plugins compatible tag[2][3] needs to be released for
their testing.

Tagging '2024.1-last' with the latest hash, also releasing
a new tag 4.0.0 (major version bump due to drop of v1.1 API
support).

[1] https://review.opendev.org/c/openstack/releases/+/963711
[2] https://docs.openstack.org/tempest/latest/stable_branch_support_policy.html
[3] https://docs.openstack.org/tempest/latest/tempest_and_plugins_compatible_version_policy.html

Change-Id: Ie150973755737d04542d044b5c0f9d4e4ed827d3
Signed-off-by: Ghanshyam Maan <gmaan@ghanshyammann.com>