Release oslo.policy for Gazpacho-1 milestone
This is a library release for oslo.policy for the Gazpacho-1 milestone.
This repository contains commits that have not been released for this
cycle yet.

If the team is ready to process with this release, please leave a +1
to indicate we should go ahead with the release.

If the team needs more time for things about to merge, or if there is
some other reason a release should not be done at this time, please
leave a -1 with a comment indicating the status. Then update the patch
with the new commit hash to use once the team is ready to do the
release.

Change-Id: Ie68a8c7cd684b3a32631a8c000a5a98da66e6261
Signed-off-by: Elod Illes <elod.illes@est.tech>
Generated-By: openstack/releases:tools/process_auto_releases.sh