oslo.cache 4.0.0
While we removed support for the mongodb backend in this release, this
had been broken for over a decade since pymongo 3.x was first released
in 2015 [1][2] and it can't have been in use with modern stacks due to
lack of modern Python 3 support in the old 2.x series. Nonetheless, we
bump the major version to be SemVer compliant.

[1] https://review.opendev.org/c/openstack/oslo.cache/+/962795
[2] https://pypi.org/project/pymongo/3.0/

Change-Id: I82b716bb4cb2f3481c6d07dec24deb6988d33d05
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>