[tool] Do not add reviewer to patch for unmaintained
Unmainained branches are not maintained by project teams, thus,
they don't need PTL or release liaison approval. So let's not
spam them.

These patches are usually the patches where we move an unmaintained
branch to End of Life. As we don't release anything out of a branch
in unmaintained state.

Change-Id: Ifac8392296a738d462070bd2bbf9f97c597369b6
Signed-off-by: Előd Illés <elod.illes@est.tech>